Clean-up, fix MVP errors, nvmrc
