+++ .nvmrc
@@ -1,0 +1,1 @@
+v17.9.0
