--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-  "cSpell.words": [
-    "savedjob"
-  ]
-}
