--- src/components/profile/ProfilePresenter.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import ProfileView from './ProfileView'
-
-const Profile = () => {
-  return (
-    <>
-      <ProfileView />
-    </>
-  )
-}
-
-export default Profile
