--- src/components/profile/ProfileView.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const ProfileView = () => {
-  return (
-    <>
-      <h1>Profile</h1>
-      <h2>CRUD for profile goes here (when we have fixed our own auth)</h2>
-    </>
-  )
-}
-
-export default ProfileView
